Create MakeMove module; Create MakeMove spec file; Initialize MakeMove in enigma
